requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
functions-framework==3.*
openai
requests
twilio
python-multipart
pinecone-client
google-generativeai
google-cloud-storage
numpy
uvicorn[standard]
websockets